booking process modulaized, booking service
global state with bookingStateProvider (activeStep)
global service with  BookingService (selectedDay, appointmentBoxes)
updated naming
