+++ lib/appointment_booking/booking/appointmentBox_widget.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/cupertino.dart';
+
+class AppointmentBox extends StatelessWidget {
+  final DateTime time;
+  final String id;
+  final VoidCallback callback;
+
+  const AppointmentBox({
+    required this.time,
+    required this.id,
+    required this.callback,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    double width = MediaQuery.of(context).size.width;
+    return CupertinoButton.filled(
+      padding: EdgeInsets.symmetric(horizontal: width * 0.09),
+      onPressed: callback,
+      child: Text(
+        ((time.hour.toString().length == 1)
+                ? "0" + time.hour.toString()
+                : time.hour.toString()) +
+            ":" +
+            ((time.minute.toString().length == 1)
+                ? "0" + time.minute.toString()
+                : time.minute.toString()),
+      ),
+    );
+  }
+}
