--- lib/appointment_booking/frage_model.dart
@@ -1,6 +1,0 @@
-class Frage {
-  String text;
-  bool isYesCorrect;
-
-  Frage(this.text, this.isYesCorrect);
-}
