+++ lib/appointment_booking/questions/question_model.dart
@@ -1,0 +1,6 @@
+class Question {
+  String text;
+  bool isYesCorrect;
+
+  Question(this.text, this.isYesCorrect);
+}
