--- lib/appointment_booking/termin_model.dart
@@ -1,16 +1,0 @@
-class Termin {
-  final String id;
-  final DateTime time;
-
-  Termin({
-    required this.id,
-    required this.time,
-  });
-
-  factory Termin.fromJson(Map<String, dynamic> json) {
-    return Termin(
-      id: json['id'],
-      time: json['time'],
-    );
-  }
-}
