+++ lib/main.dart
@@ -56,8 +56,11 @@ class MyApp extends StatelessWidget {
   }
 
   Future<bool> showOnboarding() async {
+    //Do not show onboarding in web version
     if (kIsWeb) return true;
 
+    //maybe use UserService for this
+    //load persistent data to check onboarded status
     SharedPreferences prefs = await SharedPreferences.getInstance();
     final alreadyOnboarded = prefs.getBool("alreadyOnboarded") ?? false;
