+++ lib/services/services.dart
@@ -2,10 +2,9 @@
 
 import 'dart:io';
 
+import 'booking/booking_services.dart';
 import 'firebase/firebase_service.dart';
 import 'user/user_service.dart';
-
-
 
 class Services {
   static final Services instance = Services._privateConstructor();
@@ -16,6 +15,8 @@ class Services {
     print("Init Services ...");
 
     UserService.instance;
+    // should be initialized later when booking process first starts
+    BookingService.instance;
 
     if (Platform.isIOS || Platform.isAndroid) {
       FirebaseService.instance;
