+++ test/Unittesting/frage_model_test.dart
@@ -1,12 +1,14 @@
-import '../../lib/appointment_booking/frage_model.dart';
+import 'package:blooddonation/appointment_booking/questions/question_model.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
-  test('Given frage_model Object is created when asked for isYesCorrect Then the defined isYesCorrect is put out',(){
+  test(
+      'Given frage_model Object is created when asked for isYesCorrect Then the defined isYesCorrect is put out',
+      () {
     // ARRANGE
-    Frage frage = Frage('Ist das eine Frage?', true);
+    Question question = Question('Ist das eine Frage?', true);
     // ACT
-  	var res = frage.isYesCorrect;
+    var res = question.isYesCorrect;
     // ASSERT
     expect(res, true);
   });
