+++ test/Widgettesting/frage_test.dart
@@ -1,19 +1,19 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:blooddonation/appointment_booking/frage_model.dart';
-import 'package:blooddonation/appointment_booking/frage.dart';
+import 'package:blooddonation/appointment_booking/questions/question_model.dart';
+import 'package:blooddonation/appointment_booking/questions/question_widget.dart';
 
 void main() {
   testWidgets(
       'Given the Question Widget When a Question and Function is given Then a Card with the Question text is created and two Elevated Buttons',
       (WidgetTester tester) async {
     //ARRANGE
-    Frage testfrage = Frage('Testfrage', true);
+    Question testfrage = Question('Testfrage', true);
     await tester.pumpWidget(ProviderScope(
       child: MaterialApp(
-        home: FragenWidget(
-          frage: testfrage,
+        home: QuestionWidget(
+          question: testfrage,
           nextQuestionFunc: () {
             return null;
           },
@@ -23,8 +23,8 @@ void main() {
 
     //ACT
     final questionFinder = find.text('Testfrage');
-    final elButtonFinderJ = find.text('Ja');
-    final elButtonFinderN = find.text('Nein');
+    final elButtonFinderJ = find.text('Yes');
+    final elButtonFinderN = find.text('No');
 
     //ASSERT
     expect(questionFinder, findsOneWidget);
